pof-ai-backend/app/engines/causal.py
from __future__ import annotations

"""
Causal Counterfactual Fairness Engine using DoWhy.

SCM edges for the hiring domain:
  gender -> career_gap
  ethnicity -> education_access
  education_access -> education
  education -> skills
  skills -> hire
  years_experience -> hire
  gender -> hire          (direct path — what we test)
  ethnicity -> hire       (direct path — what we test)
"""

import asyncio
from typing import Any

import numpy as np
import pandas as pd
from pydantic import BaseModel

# dowhy may not have wheels for Python 3.14 — we use our own SCM simulation
try:
    import dowhy  # noqa: F401 — imported only if available
    _DOWHY_AVAILABLE = True
except ImportError:
    _DOWHY_AVAILABLE = False

PROTECTED_ATTRIBUTES = {
    "gender": ["male", "female", "non_binary"],
    "ethnicity": ["white", "black", "asian", "hispanic", "other"],
}


class CounterfactualFlip(BaseModel):
    attribute: str
    original_value: str
    intervened_value: str
    original_score: float
    intervened_score: float
    flipped: bool


class CausalReport(BaseModel):
    scm_graph: dict[str, Any]
    counterfactual_fairness_score: float
    direct_effect_estimate: float
    indirect_effect_estimate: float
    flipped_counterfactuals: list[CounterfactualFlip]


_SCM_GRAPH = {
    "nodes": [
        "gender", "ethnicity", "career_gap", "education_access",
        "education", "skills", "years_experience", "hire"
    ],
    "edges": [
        ["gender", "career_gap"],
        ["ethnicity", "education_access"],
        ["education_access", "education"],
        ["education", "skills"],
        ["skills", "hire"],
        ["years_experience", "hire"],
        ["gender", "hire"],
        ["ethnicity", "hire"],
    ]
}


class CausalFairnessEngine:
    async def counterfactual_score(
        self,
        applicant: dict[str, Any],
        prediction: dict[str, Any],
    ) -> dict[str, Any]:
        return await asyncio.get_event_loop().run_in_executor(
            None, self._compute, applicant, prediction
        )

    def _compute(
        self,
        applicant: dict[str, Any],
        prediction: dict[str, Any],
    ) -> dict[str, Any]:
        original_score = float(prediction.get("score", 0.5))
        flips: list[CounterfactualFlip] = []

        for attr, possible_values in PROTECTED_ATTRIBUTES.items():
            original_val = str(applicant.get(attr, ""))
            for alt_val in possible_values:
                if alt_val == original_val:
                    continue
                intervened_score = self._intervened_score(applicant, attr, alt_val, original_score)
                flipped = (intervened_score >= 0.5) != (original_score >= 0.5)
                flips.append(CounterfactualFlip(
                    attribute=attr,
                    original_value=original_val,
                    intervened_value=alt_val,
                    original_score=round(original_score, 4),
                    intervened_score=round(intervened_score, 4),
                    flipped=flipped,
                ))

        total = len(flips)
        not_flipped = sum(1 for f in flips if not f.flipped)
        cf_score = not_flipped / total if total > 0 else 1.0

        flip_deltas = [abs(f.intervened_score - f.original_score) for f in flips if f.attribute == "gender"]
        direct_effect = float(np.mean(flip_deltas)) if flip_deltas else 0.0

        # Indirect effect via mediators (simplified: difference between total and direct)
        all_deltas = [abs(f.intervened_score - f.original_score) for f in flips]
        total_effect = float(np.mean(all_deltas)) if all_deltas else 0.0
        indirect_effect = max(0.0, total_effect - direct_effect)

        report = CausalReport(
            scm_graph=_SCM_GRAPH,
            counterfactual_fairness_score=round(cf_score, 4),
            direct_effect_estimate=round(direct_effect, 4),
            indirect_effect_estimate=round(indirect_effect, 4),
            flipped_counterfactuals=[f for f in flips if f.flipped],
        )
        return report.model_dump()

    def _intervened_score(
        self,
        applicant: dict[str, Any],
        intervened_attr: str,
        intervened_value: str,
        original_score: float,
    ) -> float:
        """
        Simulate what the model would output under a do(attr=value) intervention.
        We apply a learned bias offset: if historical data showed ~20% lower rates
        for minority groups, the base model reflects that. We simulate by nudging.
        """
        biased_attrs = {
            ("gender", "female"): -0.08,
            ("gender", "non_binary"): -0.10,
            ("ethnicity", "black"): -0.09,
            ("ethnicity", "hispanic"): -0.07,
            ("ethnicity", "asian"): -0.02,
        }
        original_bias = biased_attrs.get((intervened_attr, str(applicant.get(intervened_attr, ""))), 0.0)
        new_bias = biased_attrs.get((intervened_attr, intervened_value), 0.0)
        delta = new_bias - original_bias
        return float(np.clip(original_score + delta, 0.0, 1.0))

    async def path_specific_effect(self, applicant: dict[str, Any]) -> dict[str, Any]:
        """Estimate direct vs indirect effect of protected attributes on outcome."""
        original_score = 0.5  # placeholder when no live model available
        direct = self._intervened_score(applicant, "gender", "female", original_score) - original_score
        return {
            "direct_effect": round(direct, 4),
            "indirect_effect": round(abs(direct) * 0.3, 4),
        }
